added some more isComplex equations
git-svn-id: https://svn.sable.mcgill.ca/mclab/Project@2213 89b902b4-d83f-0410-b942-89f7a419ffdb

Migrated from Sable/mclab@ca88ca8225b989d1d56488684ad7ab2120529844

+++ languages/Natlab/src/natlab/tame/AdvancedTamerTool.java
@@ -22,6 +22,7 @@ import natlab.toolkits.path.FileEnvironment;
 import natlab.toolkits.path.FilePathEnvironment;
 
 public class AdvancedTamerTool {
+	
 	public IntraproceduralValueAnalysis<AggrValue<AdvancedMatrixValue>> 
     tameMatlabToSingleFunction(java.io.File mainFile, List<AggrValue<AdvancedMatrixValue>> inputValues){
 		GenericFile gFile = GenericFile.create(mainFile); //file -> generic file
@@ -84,8 +85,8 @@ public static void main(String[] args){
 
 public static List<AggrValue<AdvancedMatrixValue>> getListOfInputValues() {
 	List<PrimitiveClassReference> ls = new ArrayList<PrimitiveClassReference>(1);
-	ls.add(PrimitiveClassReference.INT8);
-	//ls.add(PrimitiveClassReference.DOUBLE);
+	//ls.add(PrimitiveClassReference.INT8);
+	ls.add(PrimitiveClassReference.DOUBLE);
 	
 	ArrayList<AggrValue<AdvancedMatrixValue>> list = new ArrayList<AggrValue<AdvancedMatrixValue>>(ls.size());
 	for (PrimitiveClassReference ref : ls){
